+++ CHANGELOG.md
@@ -2,3 +2,15 @@
 
 ### 0.0.1
 - Initial release 
+- Load API Definitions from json file.
+- Import API Definition json from Postman exported json file (Collection v2.1)
+
+### 0.0.2
+- Change private access specifier to protected  
+- Introduce properties to set default Definition `.json` and source (Postman) `.json`
+- Fix recursive merge of construction parameters with default options
+- Recursive Merge child default properties
+- Set child class's directory as rootPath 
+- Documentation generator [under-development]
+
+